+++ src/providers/calendarEventTitleFormatter.provider.ts
@@ -3,7 +3,7 @@ import { CalendarEvent } from 'calendar-utils';
 /**
  * This class is responsible for displaying all event titles within the calendar. You may override any of its methods via angulars DI to suit your requirements. For example:
  *
- * ```
+ * ```typescript
  * import { CalendarEventTitleFormatter, CalendarEvent } from 'angular-calendar';
  *
  * class CustomEventTitleFormatter extends CalendarEventTitleFormatter {