+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.12.0"></a>
+# [0.12.0](https://github.com/mattlewis92/angular-calendar/compare/v0.11.0...v0.12.0) (2017-04-05)
+
+
+### Features
+
+* **dayView:** allow the all day event template to be customised ([d542d13](https://github.com/mattlewis92/angular-calendar/commit/d542d13))
+* **dayView:** allow the hour segment template to be customised ([149b605](https://github.com/mattlewis92/angular-calendar/commit/149b605)), closes [#172](https://github.com/mattlewis92/angular-calendar/issues/172)
+
+
+
 <a name="0.11.0"></a>
 # [0.11.0](https://github.com/mattlewis92/angular-calendar/compare/v0.10.1...v0.11.0) (2017-03-29)